src/requirements.txt
Pillow==9.0.1
face-recognition==1.3.0
# opencv-python==4.5.5.62